test refactoring and uptime monitor coverage
